Fix not sending will on clean start during re-connect
If a client re-connects with a clean start, the specification
defines that this "ends" and existing sessions.

Whenever a session ends, the specification defines that
the ending session's will must be sent, if available and
still pending due to a session expiry interval or a will
delay interval that has not been reached yet.

This is achieved by sending any still pending will between
disconnecting and (re-)connecting.

Also improved method names of PendingWillMessages.