Log a warning if a will can't be sent because the session is gone
For testing purposes, assert that this never happens
because it is unexpected: We must guarantee that a
will is sent before the session is gone/cleaned up.

To detect unknown race conditions in production,
log a warning and don't throw a NullPointerException,
as this might block client connections and the
sending of other wills.